+++ HISTORY.md
@@ -789,6 +789,10 @@ domain/Todo/validators 内の各種の関数のテストコードを書いた。
 - テスト実行時の N 日前の日付について、関数で数値タプルや日付文字列を作成している。
 
 ```TypeScript
+// 補助関数
+/**
+ * [yyyy-MM-dd] 形式のローカル日付 (e.g. JST)
+ */
 function formatDate(date: Date): string {
   const y = String(date.getFullYear());
   const m = String(date.getMonth() + 1).padStart(2, '0');
@@ -797,13 +801,29 @@ function formatDate(date: Date): string {
   return `${y}-${m}-${d}`;
 }
 
-function getDateStringBefore(date: Date, days: -1 | 0 | 1 | 2): string {
+/**
+ * date から \\{days} 日前の日付を表す文字列を返す
+ *
+ * - days の小数点以下は切り捨てる。
+ * @param date 基準となる日に対応する Date オブジェクト
+ * @param days 遡りたい日数
+ * @returns [yyyy-mm-dd] 形式の \\{days} 日前の日付文字列
+ */
+export function getDateStringBefore(date: Date, days: -1 | 0 | 1 | 2): string {
   const copy = new Date(date);
   copy.setDate(copy.getDate() - days);
 
   return formatDate(copy);
 }
 
+/**
+ * date から \\{days} 日前の日付を表す数値のタプルを返す
+ *
+ * - days の小数点以下は切り捨てる。
+ * @param date 基準となる日に対応する Date オブジェクト
+ * @param days 遡りたい日数
+ * @returns [yyyy, mm, dd] の \\{days} 日前の数値タプル
+ */
 export function getDateTupleBefore(date: Date, days: -1 | 0 | 1 | 2): [number, number, number] {
   const copy = new Date(date);
   copy.setDate(copy.getDate() - days);